refactor: remove the toqito/helper package, move its modules into matrix_ops
The toqito/helper package held two cvxpy<->numpy conversion helpers used only
by matrix_ops.partial_trace and matrix_ops.partial_transpose. Move them next to
their only callers as regular matrix_ops modules and drop the standalone
helper package:

  helper/expr_as_np_array.py -> matrix_ops/expr_as_np_array.py
  helper/np_array_as_expr.py -> matrix_ops/np_array_as_expr.py

Importers and tests updated to the new paths; tests moved into
matrix_ops/tests. No logic changes.
