--- toqito/helper/__init__.py
@@ -1,4 +1,0 @@
-"""Helper is a set of modules that implements helper functions for dealing with cvxpy objects."""
-
-from toqito.helper.expr_as_np_array import expr_as_np_array
-from toqito.helper.np_array_as_expr import np_array_as_expr
