+++ toqito/matrix_ops/partial_trace.py
@@ -4,7 +4,8 @@ import numpy as np
 from cvxpy.expressions.expression import Expression
 from cvxpy.expressions.variable import Variable
 
-from toqito.helper import expr_as_np_array, np_array_as_expr
+from toqito.matrix_ops.expr_as_np_array import expr_as_np_array
+from toqito.matrix_ops.np_array_as_expr import np_array_as_expr
 from toqito.perms import permute_systems
